+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix `accessibilityValidation`
 - remove the need for `experimentalRunEvents` in cypress >= `6.7.0`
 
 ## 3.21.0 - 2021/4/22