+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
+
+## 3.21.1 - 2021/5/12
+
 - fix `accessibilityValidation`
 - remove the need for `experimentalRunEvents` in cypress >= `6.7.0`
+- updated to @applitools/visual-grid-client@15.8.4 (from 15.8.0)
 
 ## 3.21.0 - 2021/4/22